refactor(parquet-datasource): split sink and schema_coercion out of file_format.rs
Pure code motion, no behavior change. `file_format.rs` had grown to
~2,000 LOC bundling several responsibilities; this extracts two of
them into focused modules (file_format.rs drops to ~660 LOC):

- `sink.rs` — `ParquetSink` and the parallel-write machinery
  (`column_serializer_task`, `spawn_column_parallel_row_group_writer`,
  `output_single_parquet_file_parallelized`,
  `concatenate_parallel_row_groups`, etc.).
- `schema_coercion.rs` — the Arrow-schema coercion utilities
  (`apply_file_schema_type_coercions`, `coerce_int96_to_resolution`,
  `coerce_file_schema_to_view_type`, `coerce_file_schema_to_string_type`,
  `transform_schema_to_view`, `transform_binary_to_string`,
  `field_with_new_type`) and their tests.

No public API change. Every previously-public item is still reachable
at the same path: the crate root re-exports `sink::ParquetSink` and the
`schema_coercion::*` functions, and the historical
`file_format::ParquetSink` path is preserved via `pub use`
(datafusion-proto depends on it).

Split out of #22156.
