+++ datafusion/datasource-parquet/src/mod.rs
@@ -34,6 +34,8 @@ mod push_decoder;
 mod reader;
 mod row_filter;
 mod row_group_filter;
+pub mod schema_coercion;
+pub mod sink;
 mod sort;
 pub mod source;
 mod supported_predicates;
@@ -47,4 +49,11 @@ pub use reader::*; // Expose so downstream crates can use it
 pub use row_filter::build_row_filter;
 pub use row_filter::can_expr_be_pushed_down_with_schemas;
 pub use row_group_filter::RowGroupAccessPlanFilter;
+#[expect(deprecated)]
+pub use schema_coercion::{
+    apply_file_schema_type_coercions, coerce_file_schema_to_string_type,
+    coerce_file_schema_to_view_type, coerce_int96_to_resolution,
+    transform_binary_to_string, transform_schema_to_view,
+};
+pub use sink::ParquetSink;
 pub use writer::plan_to_parquet;
